requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
feedparser==6.0.10
beautifulsoup4==4.12.2
schedule==1.2.0
requests==2.31.0
python-dotenv==1.0.0
smtplib-ssl==1.0.0
urllib3==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1 